+++ Runtime/MXTimeline/MXClipBehaviour.cs
@@ -4,12 +4,15 @@ namespace Voxell.MotionGFX
 {
   public class MXClipBehaviour : PlayableBehaviour
   {
-    public MXScene clipGroup;
+    public MXScene scene;
 
-    public override void ProcessFrame(Playable playable, FrameData info, object playerData)
-    {
-      float playableTime = (float) playable.GetTime();
-      clipGroup.Evaluate(playableTime);
-    }
+    // public override void ProcessFrame(Playable playable, FrameData info, object playerData)
+    // {
+    //   if (scene == null) return;
+
+    //   float playableTime = (float) playable.GetTime();
+    //   ISeqHolder seqHolder = scene as ISeqHolder;
+    //   seqHolder.Evaluate(playableTime);
+    // }
   }
 }